[Data] Move database files to local package
