+++ app/src/main/java/com/github/raziu75/tricount/data/local/entity/TripEntity.kt
@@ -1,4 +1,4 @@
-package com.github.raziu75.tricount.data.entity
+package com.github.raziu75.tricount.data.local.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
