+++ app/src/main/java/com/github/raziu75/tricount/data/local/entity/relation/ParticipantWithTransactions.kt
@@ -1,10 +1,10 @@
-package com.github.raziu75.tricount.data.entity.relation
+package com.github.raziu75.tricount.data.local.entity.relation
 
 import androidx.room.Embedded
 import androidx.room.Junction
 import androidx.room.Relation
-import com.github.raziu75.tricount.data.entity.TransactionEntity
-import com.github.raziu75.tricount.data.entity.ParticipantEntity
+import com.github.raziu75.tricount.data.local.entity.TransactionEntity
+import com.github.raziu75.tricount.data.local.entity.ParticipantEntity
 
 data class ParticipantWithTransactions(
     @Embedded val participants: ParticipantEntity,
