+++ app/src/main/java/com/github/raziu75/tricount/data/local/entity/relation/TripWithTransactions.kt
@@ -1,11 +1,9 @@
-package com.github.raziu75.tricount.data.entity.relation
+package com.github.raziu75.tricount.data.local.entity.relation
 
 import androidx.room.Embedded
-import androidx.room.Junction
 import androidx.room.Relation
-import com.github.raziu75.tricount.data.entity.TransactionEntity
-import com.github.raziu75.tricount.data.entity.ParticipantEntity
-import com.github.raziu75.tricount.data.entity.TripEntity
+import com.github.raziu75.tricount.data.local.entity.TransactionEntity
+import com.github.raziu75.tricount.data.local.entity.TripEntity
 
 data class TripWithTransactions(
     @Embedded val trip: TripEntity,
